main: Make sure we switch to the DFU interface
"usb_set_altinerface(_rt_dif.dev_handle, 0)" call seems to be necessary
to actually switch to the DFU interface when the USB-Device still uses
its default interface.
usb_claim_interface() doesn't send anything over the bus, but the
device needs to know when to serve DFU-Requests (like the following
dfu_get_status() request).
it works without it on devices where the DFU-Interface == 0 or the only
DFU interface available.
It doesn't work if a device offers a DFU-Device Descriptor in normal
runtime mode (to serve appDETACH requests).

For OM devices we did not hit the problem, but we have no problems with it
either.

Reported-by: Daniel Hiepler <nftlte@googlemail.com>

git-svn-id: https://svn.openmoko.org/trunk/src/host/dfu-util@5934 99fdad57-331a-0410-800a-d7fa5415bdb3