main: Look for DFU functional descriptor among all descriptors
Some devices are not able to return non-standard type descriptors when
asked explicitly, but will include them within the configuration
hierarchy returned when requesting the full configuration descriptor.

Section 9.4.3 of the USB 1.1 specification does not dictate support
for requesting a "DFU functional descriptor" type, as I read it:

"The standard request to a device supports three types of descriptors:
DEVICE, CONFIGURATION, and STRING." and further "Class-specific and/or
vendor-specific descriptors follow the standard descriptors they
extend or modify."

This patch first looks for the functional descriptor in the full set of
descriptors. If it is not found, it will do the direct type=0x21 request
like before.

The patch also replaces a magic 0x21 with its more verbose DEFINE.