main: New default (1024) and clamping of transfer size
DFU dictates that the effective transfer size is chosen from
bMaxPacketSIze0 (often 64) up to the real wTransferSize in the device.
If we can not detect the real value of wTransferSize, lower guesses can
only be safer. The previous default was the host kernel page size which
is the higher limit on the host side (on Linux). This would usually
translate into 4096, which possibly is too high for some devices.

1024 bytes seems like a reasonable trade-off for a default size, when
one is needed. At any rate, it should really be determined from the DFU
functional descriptor, so this default should not matter much.