quirks: Add OpenOCD to the poll timeout quirk table.
This fixes a regression we introduced with the poll timeout handling. The
sam7dfu bootloader on the OpenPCD does not set the value and therefor acts like
the Openmoko devices. Use the same quirk for it.
Thanks to Bitmanufaktur for sending me a device to keep the support in dfu-util
working.