dfu_file.c: Add parsing of DFU file suffix
The new dfu_file structure is used to keep track of the download (or
upload) file. parse_dfu_suffix() will parse an (already opened) file and
populate the struct members corresponding to the DFU suffix fields.

The commit also includes some verification of the suffix fields,
but at this point only warns about mismatches, with one exception:
If there is a valid DFU suffix, it must have DFU revision 1.0,
otherwise we abort the download.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>