main: Count matching interface/altsetting
Check that the users filter options (if any) fully specify one and only
one interface/altsetting by using count_matching_dfu_if(). This fixes
for instance the following scenario: If you had one interface with two
altsettings, the program would stop and complain that you had to choose
one using -i or -a. Then if you chose -i 0 you actually did not sort out
the ambiguity, however the program would go on to select a=0 for you...

Also use a get_matching_dfu_if() to make sure that the dif structure is
fully populated, so that its description can be printed out for
verification, and so that the download/upload functions has access to
all information through the dif structure.