main: Rename MAX_STR_LEN and bump it to 253
The way we retrieve these strings we just have to make sure the buffer
is big enough to not truncate the string.

A USB descriptor is limited to 255 bytes since its bLength field is
one byte wide. So it can theoretically contain 253 characters if
UTF-8 encoding is used. UTF-16 encoding has been mandated since 2005
but we can afford those extra bytes just in case.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>