main: Retrieve run-time and DFU mode functional descriptor
Previously we only retrieved the DFU mode descriptor and only if the
user had not specified the transfer size option.

We will try to get the run-time mode descriptor as cached by libusb,
without asking the device (which would upset e.g. the Freerunner).
This way we can determine e.g. the DFU version before dealing more
with the device.

After device transition to DFU mode we will retrieve the DFU mode
descriptor, first by trying to get it cached from libusb, and if that
fails, by asking the device. If that also fails, we'll guess wisely.