device-logs: Verbose lsusb logs of DFU enabled devices.
This picks up an idea from Tormod to keep lsusb -v logs for DFU enabled devices
we own. This should make our life easier when developing code to avoid
regressions more easily.

If you have a device that is supported by dfu-util and not listed here please
prepare a log and a description for it and send it over.