main: Honor the bitWillDetach flag (DFU 1.1)
If the device has the bitWillDetach set it will detach and reattach
itself to the USB bus after receiving the DETACH request, and the
host computer should not issue a USB reset.

Also make sure the descriptor structures containing these
flags are initialized to zero, for the case where the
descriptor can not be retrieved from the device.