main: List alternate interfaces in DFU mode
in the case there are several and the user did not specify one.

This can be useful if the device does not expose alternate
interfaces in run-time mode.

Also adjusted a warning message to clarify that we do not set
the alternate interface in run-time mode, only in DFU mode.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>