main: Update copyright banner
Stefan and I have contributed substantially to the whole
code base, so add us both as authors.

Spell out "Copyright" since "(C)" has no legal meaning.
IANAL, YMMV :)

The warning about DFU version 1.0 only is not necessary,
we support 1.1 as well as 1.0 (and 1.1a). And there are no
other versions of the standard.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>