Port file handling to stdio streams
Thanks go to Satz Klauer, some of this is inspired by his Windows
patch.

Note that the "delete DFU suffix" operation of the dfu-suffix tool still
uses the POSIX ftruncate() to truncate a file, so this part is less
portable. We use autotools to check for availability of ftruncate().
This might not be available in all win32 environments, but we recommend
using MinGW for Windows, where this works fine.

In main.c, for opening the uploading file, the previously used open()
with O_CREATE|O_EXCL flags does not have a straight equivalent in stdio.
glibc has an fopen() "wx" mode equivalent to the posix open()
O_CREAT|O_EXCL mode, but to make this more portable we will instead open
for appending, then check if we are at the beginning of the file.

So the only difference is that this will check for a non-empty
file whereas the old open() would check for an existing file.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>