configure.ac: Remove unnecessary tests
Some of these checks are not necessary and we do not use
their corresponding HAVE_XXX in the code.

The AC_FUNC_MALLOC compliance test is not needed on our
supported(?) targets, and it also does the wrong thing
when cross-compiling for Windows using mingw32.

Also add -i (install missing auxiliary files) to autoreconf
in autogen.sh to generate e.g. m4 files.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>