dfuse: Add "force" modifier
The "force" modifier can be added to the --dfuse-address
argument, separated by a colon.

It is meant as an expert option, and its use and abuse can
cause all kinds of trouble, brick your device, etc.

For now, it allows uploading from a memory location not
advertised in the memory map. The bootloader may or may not
deal with this.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>