dfuse: Add support for sending DFU_LEAVE request
The "leave" modifier can be added to the --dfuse-address
argument, separated by a colon.

It tells dfu-util to send a DFU_LEAVE request after
downloading/uploading, which is meant to cause the device
to leave DFU mode and start the normal application, often
by resetting itself.

DFU Leave operation is described in ST document AN3156 section 5.5.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>