dfuse: Add support for Read Unprotect command
Thanks to Uwe Bonnes for writing an initial patch.

The command is described in ST document AN3156 section 5.4

The "unprotect" modifier can be added to the --dfuse-address argument,
separated by a colon. Since this command will erase the whole flash
memory, it must be combined with the "force" option to take effect.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>