configure.ac: Check for libusb 1.0 in system location before using pkg-config
Fixes libusb 1.0 detection on FreeBSD where the library is called simply
"libusb" and resides in the system default location.

Note that pkg-config (or more precisely, its aclocal/pkg.m4) must be
installed when running autogen.sh, but pkg-config is not needed on
FreeBSD when running configure (which is included in release tarballs).

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>