dfuse: More verbose output from special command download
For instance the Erase operation will fail if the pages are write
protected or the address is not allowed. We could be more helpful
decoding the returned status but at least we print the status codes
so that the user can look them up in the documention.

Another example is the mass-erase command on the STM32F4-Discovery
board which seems to perform the erase correctly, but returns an
unknown error in the second get_status request.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>